Splits files that can be reused by the app extension
